requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.16.0
cryptography==42.0.7
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
ecdsa==0.19.0
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
python-jose==3.3.0
pytz==2024.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.0
tzdata==2024.1
whitenoise==6.6.0
